Abstract the runtime used in runtime tests so that it's not hardcoded as Spin
